+++ apps/videos/serializers/__init__.py
@@ -1,0 +1,3 @@
+from .video_serializer import VideoSerializer
+
+__all__ = ['VideoSerializer']
