+++ apps/videos/views/__init__.py
@@ -1,0 +1,7 @@
+from .video_crud_views import VideoUploadView
+# from .video_creation_view import VideoCreationView  # 추후 구현 예정
+
+__all__ = [
+    'VideoUploadView',
+    # 'VideoCreationView',  # 추후 구현 예정
+]
