+++ apps/videos/views/video_creation_view.py
@@ -1,0 +1,29 @@
+from rest_framework.views import APIView
+from rest_framework.response import Response
+from rest_framework import status
+from rest_framework.permissions import AllowAny
+from ..models import Video
+from ..serializers import VideoSerializer
+from drf_yasg.utils import swagger_auto_schema
+from drf_yasg import openapi
+
+
+# TODO: 작품 기반 영상 생성 API 구현 예정
+#
+# class VideoCreationView(APIView):
+#     """
+#     작품 기반 영상 생성 API
+#     
+#     작품 정보를 입력받아 자동으로 영상을 생성하는 기능
+#     예: 작품명, 아티스트, 장소 정보 등을 기반으로 영상 템플릿 생성
+#     """
+#     permission_classes = [AllowAny]
+#     
+#     @swagger_auto_schema(
+#         operation_description="작품 정보를 기반으로 영상을 자동 생성합니다",
+#         # request_body 및 responses 추후 정의
+#     )
+#     def post(self, request):
+#         """작품 기반 영상 생성"""
+#         # 구현 예정
+#         pass
